% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHyperConnection.R
\name{dbAttachDatabase,HyperConnection,character,character-method}
\alias{dbAttachDatabase,HyperConnection,character,character-method}
\title{Add a Hyper database to the current Hyper process.}
\usage{
\S4method{dbAttachDatabase}{HyperConnection,character,character}(conn, db_file, name = NA_character_, ...)
}
\description{
Add a Hyper database to the current Hyper process.
}
