% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHyperConnection.R
\name{dbDetachDatabase,HyperConnection,character-method}
\alias{dbDetachDatabase,HyperConnection,character-method}
\title{Remove a Hyper database from the current Hyper process.}
\usage{
\S4method{dbDetachDatabase}{HyperConnection,character}(conn, name, ...)
}
\description{
Remove a Hyper database from the current Hyper process.
}
