% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHyperConnection.R
\name{dbDisconnect,HyperConnection-method}
\alias{dbDisconnect,HyperConnection-method}
\title{Terminate a Hyper Connection.}
\usage{
\S4method{dbDisconnect}{HyperConnection}(conn)
}
\arguments{
\item{HyperConnect}{}
}
\description{
Terminate a Hyper Connection.
}
