% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHyperResult.R
\name{dbFetch,HyperResult-method}
\alias{dbFetch,HyperResult-method}
\title{Retrieve records from Hyper query}
\usage{
\S4method{dbFetch}{HyperResult}(res, n = -1, ...)
}
\description{
Retrieve records from Hyper query
}
