% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHyperResult.R
\name{dbGetRowsAffected,HyperResult-method}
\alias{dbGetRowsAffected,HyperResult-method}
\title{Retrieve records from Hyper query}
\usage{
\S4method{dbGetRowsAffected}{HyperResult}(res, ...)
}
\description{
Retrieve records from Hyper query
}
