% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHyperConnection.R
\name{dbSendQuery,HyperConnection-method}
\alias{dbSendQuery,HyperConnection-method}
\title{Send a query to Hyper.}
\usage{
\S4method{dbSendQuery}{HyperConnection}(conn, statement, ...)
}
\description{
Send a query to Hyper.
}
