% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHyperDriver.R
\docType{class}
\name{HyperDriver-class}
\alias{HyperDriver-class}
\alias{dbUnloadDriver,HyperDriver-method}
\alias{show,HyperDriver-method}
\alias{Hyper}
\alias{dbConnect,HyperDriver-method}
\alias{dbGetInfo,HyperDriver-method}
\title{Driver for Hyper database.}
\usage{
\S4method{dbUnloadDriver}{HyperDriver}(drv, ...)

\S4method{show}{HyperDriver}(object)

Hyper()

\S4method{dbConnect}{HyperDriver}(drv, telemetry = FALSE, database = list(), bigint = "numeric", ...)

\S4method{dbGetInfo}{HyperDriver}(dbObj, ...)
}
\arguments{
\item{drv}{An object created by \code{Hyper()}}

\item{HyperDriver}{}
}
\description{
Driver for Hyper database.
}
\keyword{internal}
