% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{in_database}
\alias{in_database}
\title{Refer to a table in a specific database.}
\usage{
in_database(database, schema = "public", table)
}
\arguments{
\item{table}{}
}
\description{
Refer to a table in a specific database.
}
